Add adder network and signed multiplier with post-normalization net

+++ stick_emulator/networks/functional/linear_combinator.py
@@ -13,7 +13,7 @@ class LinearCombinatorNetwork(SpikingNetworkModule):
         self,
         encoder: DataEncoder,
         N: int,
-        coeff: list[int],
+        coeff: list[float],
         module_name: Optional[str] = None,
     ):
         super().__init__(module_name)
